WIP Skeleton of separation between gazebo and ignition plugins
