+++ building_sim_plugins/building_sim_common/package.xml
@@ -1,0 +1,33 @@
+<package format="3">
+  <name>building_sim_common</name>
+  <version>0.0.0</version>
+  <description>
+    Gazebo plugins so that buildings generated by building_map_tools can
+    talk to the ROS 2-based systems in the rmf_core package.
+  </description>
+  <author>Morgan Quigley</author>
+  <author>Michael Grey</author>
+  <author>Roselle Carmen</author>
+  <author>Aaron Chong</author>
+  <author>Brandon Ong</author>
+  <maintainer email="morgan@openrobotics.org">Morgan Quigley</maintainer>
+  <license>Apache 2.0</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <depend>rclcpp</depend>
+  <depend>gazebo_ros</depend>
+  <depend>gazebo_msgs</depend>
+  <depend>gazebo_dev</depend>
+  <depend>rmf_fleet_msgs</depend>
+  <depend>rmf_door_msgs</depend>
+  <depend>tf2_ros</depend>
+  <depend>geometry_msgs</depend>
+  <depend>std_msgs</depend>
+  <depend>std_srvs</depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+
+</package>
