+++ building_sim_plugins/building_sim_gazebo/package.xml
@@ -25,6 +25,7 @@
   <depend>geometry_msgs</depend>
   <depend>std_msgs</depend>
   <depend>std_srvs</depend>
+  <depend>building_sim_common</depend>
 
   <export>
     <build_type>ament_cmake</build_type>
