+++ building_sim_plugins/building_sim_gazebo/src/door.cpp
@@ -9,7 +9,7 @@
 #include <rmf_door_msgs/msg/door_state.hpp>
 #include <rmf_door_msgs/msg/door_request.hpp>
 
-#include "utils.hpp"
+#include <building_sim_common/utils.hpp>
 
 namespace building_gazebo_plugins {
