+++ building_sim_plugins/building_sim_gazebo/src/toggle_floors.cpp
